create or update data package with EDI API

create_package creates or updates a data package in the EDI data repository.

Usage

create_package(
  identifier = NULL,
  scope = NULL,
  environment = "staging",
  update = FALSE
)

Arguments

identifier

(integer) Unquoted package identifier integer

scope

(character) Quoted package scope (e.g., "knb-lter-cap", "edi")

environment

(character) Quoted name of the EDI environment where the data package should be created, can be: production, staging (default), or development.

update

(logical) Boolean value indicating if the data package already exists and should be updated.

Note

create_package will look for a config.yaml file to estalish the package scope and identifier. Alternatively, these parameters can be provided by the user but, in that case, both parameters must be provided.

The user must have established a connection to the EDI API (see EDIutils::login) in the current R session.