write a project config.yaml file

write_config writes a template config.yaml file to the working directory (default) or supplied path.

Usage

write_config(scope = "knb-lter-cap", identifier, path = ".", overwrite = FALSE)

Arguments

scope

(character) Quoted name of the package scope (e.g., "edi"). The default is "knb-lter-cap".

identifier

(integer) Data package identifier (number).

path

(character) Path to where the config file will be written. Defaults to the current directory.

overwrite

(logical) Logical indicating if an existing config file in the target directory should be overwritten.

Details

A template config.yaml file is written to the working directory or specified location. The function requires that the package scope and number (e.g., "edi", 521) are provided. Package identifiers include a verion number; the function uses a default value of version 1 but an alternate version number can be specified.

Note

The function expects specifically a 3-digit package identifier (number).