create_spatialRaster

create_spatialRaster generates a EML entity of type spatialRaster

create_spatialRaster(rasterFile, description, epsgProjection,
  emlProjection, rasterValueDescription, rasterValueUnits,
  zipFiles = FALSE,
  baseURL = "https://data.gios.asu.edu/datasets/cap/")

Arguments

rasterFile
Quoted full path to raster file.

description
Description of the raster.

epsgProjection
EPSG numeric code of raster's coordinate reference system

emlProjection
EML-compliant refence to raster's coordinate reference system

rasterValueDescription
Description of raster values

rasterValueUnits
Raster value units

zipFiles
Logical indicating whether spatial raster entity should be constructed from a single raster file (FALSE, default) or entire directory (TRUE)

baseURL
(optional) The base path of the web-accessible location of the data file; the name of the resulting file will be passed to the base path to generate a web-resolvable file path.

Value

EML spatial data object is returned. Additionally, the spatial data file is renamed with the project id + base file name + md5sum + file extension (zip in the case when multiple files are aggregated).

Details

a spatialRaster entity is created from a single data file (e.g., CAP_1985.img) or a collection of related files (e.g., CAP_1985.img, CAP_1985.img.aux.xml). In the case of multiple files, all files in the parent directory where the raster file is located are aggregated into a single compressed (zipped) file. In all cases, the resulting entity is renamed with the project id + base file name + md5sum + file extension (zip in the case when multiple files are aggregated).

Note

create_spatialRaster will look for a project id in the working environment; this parameter is not passed to the function and it must exist.

Examples